exit the combine-accent layer by pressing the activator again.
Esc and Backspace should also do this, but still send something to the computer, so no good.

super important when you realize a typo before hitting the second key. Because KC_NO in QMK is a lie – it seems to behave just like KC_TRNS, but only on OSL, not OSM.